Extract TasksColumn in its own component and split TestTag of recently split components
