+++ tasks-app-shared/src/commonMain/kotlin/net/opatry/tasks/app/ui/screen/taskListsPane.kt
@@ -24,6 +24,7 @@ package net.opatry.tasks.app.ui.screen
 
 import CircleFadingPlus
 import LucideIcons
+import androidx.annotation.VisibleForTesting
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -57,7 +58,8 @@ import net.opatry.tasks.resources.task_lists_screen_add_task_list_cta
 import org.jetbrains.compose.resources.stringResource
 
 
-object TaskListsPaneTestTag {
+@VisibleForTesting
+internal object TaskListsPaneTestTag {
     const val NEW_TASK_LIST_BUTTON = "NEW_TASK_LIST_BUTTON"
     const val TASK_LIST_ROW = "TASK_LIST_ROW"
 }
