+++ tasks-app-shared/src/commonMain/kotlin/net/opatry/tasks/app/ui/screen/tasksPane.kt
@@ -23,30 +23,20 @@
 package net.opatry.tasks.app.ui.screen
 
 import CalendarDays
-import CheckCheck
-import ChevronDown
-import ChevronRight
-import CircleOff
 import ListPlus
 import LucideIcons
 import NotepadText
 import Plus
 import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
@@ -84,8 +74,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
@@ -95,23 +83,14 @@ import kotlinx.datetime.TimeZone
 import kotlinx.datetime.atStartOfDayIn
 import kotlinx.datetime.toLocalDateTime
 import net.opatry.tasks.app.ui.TaskListsViewModel
-import net.opatry.tasks.app.ui.component.CompletedTaskRow
 import net.opatry.tasks.app.ui.component.EditTextDialog
-import net.opatry.tasks.app.ui.component.EmptyState
-import net.opatry.tasks.app.ui.component.RemainingTaskRow
-import net.opatry.tasks.app.ui.component.RowWithIcon
-import net.opatry.tasks.app.ui.component.TaskAction
 import net.opatry.tasks.app.ui.component.TaskListEditMenuAction
 import net.opatry.tasks.app.ui.component.TaskListTopAppBar
+import net.opatry.tasks.app.ui.component.TasksColumn
 import net.opatry.tasks.app.ui.component.toColor
 import net.opatry.tasks.app.ui.component.toLabel
-import net.opatry.tasks.app.ui.model.DateRange
 import net.opatry.tasks.app.ui.model.TaskListUIModel
 import net.opatry.tasks.app.ui.model.TaskUIModel
-import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.ALL_COMPLETE_EMPTY_STATE
-import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.COMPLETED_TASKS_TOGGLE
-import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.COMPLETED_TASKS_TOGGLE_LABEL
-import net.opatry.tasks.data.TaskListSorting
 import net.opatry.tasks.resources.Res
 import net.opatry.tasks.resources.dialog_cancel
 import net.opatry.tasks.resources.task_due_date_update_cta
@@ -126,12 +105,9 @@ import net.opatry.tasks.resources.task_editor_sheet_title_field_empty_error
 import net.opatry.tasks.resources.task_editor_sheet_title_field_label
 import net.opatry.tasks.resources.task_editor_sheet_title_field_placeholder
 import net.opatry.tasks.resources.task_editor_sheet_validate
-import net.opatry.tasks.resources.task_list_pane_all_tasks_complete_desc
-import net.opatry.tasks.resources.task_list_pane_all_tasks_complete_title
 import net.opatry.tasks.resources.task_list_pane_clear_completed_confirm_dialog_confirm
 import net.opatry.tasks.resources.task_list_pane_clear_completed_confirm_dialog_message
 import net.opatry.tasks.resources.task_list_pane_clear_completed_confirm_dialog_title
-import net.opatry.tasks.resources.task_list_pane_completed_section_title_with_count
 import net.opatry.tasks.resources.task_list_pane_delete_list_confirm_dialog_confirm
 import net.opatry.tasks.resources.task_list_pane_delete_list_confirm_dialog_message
 import net.opatry.tasks.resources.task_list_pane_delete_list_confirm_dialog_title
@@ -140,27 +116,9 @@ import net.opatry.tasks.resources.task_list_pane_rename_dialog_title
 import net.opatry.tasks.resources.task_list_pane_task_deleted_snackbar
 import net.opatry.tasks.resources.task_list_pane_task_deleted_undo_snackbar
 import net.opatry.tasks.resources.task_list_pane_task_restored_snackbar
-import net.opatry.tasks.resources.task_lists_screen_empty_list_desc
-import net.opatry.tasks.resources.task_lists_screen_empty_list_title
 import net.opatry.tasks.resources.task_menu_move_to_new_list_create_task_list_dialog_confirm
 import net.opatry.tasks.resources.task_menu_move_to_new_list_create_task_list_dialog_title
 import org.jetbrains.compose.resources.stringResource
-
-object TaskListPaneTestTag {
-    const val ALL_COMPLETE_EMPTY_STATE = "ALL_COMPLETE_EMPTY_STATE"
-    const val REMAINING_TASK_ROW = "REMAINING_TASK_ROW"
-    const val REMAINING_TASK_ICON = "REMAINING_TASK_ICON"
-    const val REMAINING_TASK_NOTES = "REMAINING_TASK_NOTES"
-    const val REMAINING_TASK_DUE_DATE_CHIP = "REMAINING_TASK_DUE_DATE_CHIP"
-    const val REMAINING_TASK_MENU_ICON = "REMAINING_TASK_MENU_ICON"
-    const val COMPLETED_TASK_ROW = "COMPLETED_TASK_ROW"
-    const val COMPLETED_TASK_ICON = "COMPLETED_TASK_ICON"
-    const val COMPLETED_TASK_NOTES = "COMPLETED_TASK_NOTES"
-    const val COMPLETED_TASK_COMPLETION_DATE = "COMPLETED_TASK_COMPLETION_DATE"
-    const val COMPLETED_TASK_DELETE_ICON = "COMPLETED_TASK_DELETE_ICON"
-    const val COMPLETED_TASKS_TOGGLE = "COMPLETED_TASKS_TOGGLE"
-    const val COMPLETED_TASKS_TOGGLE_LABEL = "COMPLETED_TASKS_TOGGLE_LABEL"
-}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -241,46 +199,36 @@ fun TaskListDetail(
         }
     ) { innerPadding ->
         Box(Modifier.padding(innerPadding)) {
-            if (taskList.isEmpty) {
-                // TODO SVG undraw.co illustration `files/undraw_to_do_list_re_9nt7.svg`
-                EmptyState(
-                    icon = LucideIcons.CircleOff,
-                    title = stringResource(Res.string.task_lists_screen_empty_list_title),
-                    description = stringResource(Res.string.task_lists_screen_empty_list_desc),
-                    modifier = Modifier.fillMaxSize(),
-                )
-            } else {
-                TasksColumn(
-                    taskLists,
-                    taskList,
-                    onToggleCompletionState = { viewModel.toggleTaskCompletionState(it.id) },
-                    onEditTask = {
-                        taskOfInterest = it
-                        showEditTaskSheet = true
-                    },
-                    onUpdateDueDate = {
-                        taskOfInterest = it
-                        showDatePickerDialog = true
-                    },
-                    onNewSubTask = {
-                        taskOfInterest = it
-                        showNewSubTaskSheet = true
-                    },
-                    onUnindent = { viewModel.unindentTask(it.id) },
-                    onIndent = { viewModel.indentTask(it.id) },
-                    onMoveToTop = { viewModel.moveToTop(it.id) },
-                    onMoveToList = { task, taskList -> viewModel.moveToList(task.id, taskList.id) },
-                    onMoveToNewList = {
-                        taskOfInterest = it
-                        showNewTaskListAlert = true
-                    },
-                    onDeleteTask = {
-                        taskOfInterest = it
-                        showUndoTaskDeletionSnackbar = true
-                        viewModel.deleteTask(it.id)
-                    },
-                )
-            }
+            TasksColumn(
+                taskLists,
+                taskList,
+                onToggleCompletionState = { viewModel.toggleTaskCompletionState(it.id) },
+                onEditTask = {
+                    taskOfInterest = it
+                    showEditTaskSheet = true
+                },
+                onUpdateDueDate = {
+                    taskOfInterest = it
+                    showDatePickerDialog = true
+                },
+                onNewSubTask = {
+                    taskOfInterest = it
+                    showNewSubTaskSheet = true
+                },
+                onUnindent = { viewModel.unindentTask(it.id) },
+                onIndent = { viewModel.indentTask(it.id) },
+                onMoveToTop = { viewModel.moveToTop(it.id) },
+                onMoveToList = { task, taskList -> viewModel.moveToList(task.id, taskList.id) },
+                onMoveToNewList = {
+                    taskOfInterest = it
+                    showNewTaskListAlert = true
+                },
+                onDeleteTask = {
+                    taskOfInterest = it
+                    showUndoTaskDeletionSnackbar = true
+                    viewModel.deleteTask(it.id)
+                },
+            )
         }
     }
 
@@ -589,140 +537,3 @@ fun TaskListDetail(
         )
     }
 }
-
-@Composable
-fun TasksColumn(
-    taskLists: List<TaskListUIModel>,
-    taskList: TaskListUIModel,
-    onToggleCompletionState: (TaskUIModel) -> Unit = {},
-    onEditTask: (TaskUIModel) -> Unit = {},
-    onUpdateDueDate: (TaskUIModel) -> Unit = {},
-    onNewSubTask: (TaskUIModel) -> Unit = {},
-    onUnindent: (TaskUIModel) -> Unit = {},
-    onIndent: (TaskUIModel) -> Unit = {},
-    onMoveToTop: (TaskUIModel) -> Unit = {},
-    onMoveToList: (TaskUIModel, TaskListUIModel) -> Unit = { _, _ -> },
-    onMoveToNewList: (TaskUIModel) -> Unit = {},
-    onDeleteTask: (TaskUIModel) -> Unit = {},
-) {
-    var showCompleted by remember(taskList.id) { mutableStateOf(false) }
-
-    LazyColumn(
-        contentPadding = PaddingValues(vertical = 8.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        if (taskList.isEmptyRemainingTasksVisible) {
-            item(key = "all_tasks_complete") {
-                EmptyState(
-                    icon = LucideIcons.CheckCheck,
-                    title = stringResource(Res.string.task_list_pane_all_tasks_complete_title),
-                    description = stringResource(Res.string.task_list_pane_all_tasks_complete_desc),
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 24.dp)
-                        .testTag(ALL_COMPLETE_EMPTY_STATE),
-                )
-            }
-        }
-
-        taskList.remainingTasks.forEach { (dateRange, tasks) ->
-            if (dateRange != null) {
-                stickyHeader(key = dateRange.key) {
-                    Box(
-                        Modifier
-                            .fillMaxWidth()
-                            .background(MaterialTheme.colorScheme.background)
-                            .padding(horizontal = 12.dp, vertical = 8.dp)
-                    ) {
-                        Text(
-                            dateRange.toLabel(sectionLabel = true),
-                            style = MaterialTheme.typography.titleSmall,
-                            color = dateRange.toColor(),
-                        )
-                    }
-                }
-            }
-            items(tasks, key = { it.id.value }) { task ->
-                RemainingTaskRow(
-                    taskLists,
-                    task,
-                    // TODO could come from the UI mapper/UI state
-                    showDate = when {
-                        taskList.sorting == TaskListSorting.Manual -> true
-                        taskList.sorting == TaskListSorting.Title -> true
-                        dateRange is DateRange.Overdue -> true
-                        else -> false
-                    }
-                ) { action ->
-                    when (action) {
-                        TaskAction.ToggleCompletion -> onToggleCompletionState(task)
-                        TaskAction.Edit -> onEditTask(task)
-                        TaskAction.UpdateDueDate -> onUpdateDueDate(task)
-                        TaskAction.AddSubTask -> onNewSubTask(task)
-                        TaskAction.Unindent -> onUnindent(task)
-                        TaskAction.Indent -> onIndent(task)
-                        TaskAction.MoveToTop -> onMoveToTop(task)
-                        is TaskAction.MoveToList -> onMoveToList(task, action.targetParentList)
-                        TaskAction.MoveToNewList -> onMoveToNewList(task)
-                        TaskAction.Delete -> onDeleteTask(task)
-                    }
-                }
-            }
-        }
-
-        if (taskList.hasCompletedTasks) {
-            stickyHeader(key = "completed") {
-                Box(
-                    Modifier
-                        .clip(MaterialTheme.shapes.large)
-                        .fillMaxWidth()
-                        .clickable { showCompleted = !showCompleted }
-                        .background(MaterialTheme.colorScheme.background)
-                        .padding(horizontal = 12.dp, vertical = 8.dp)
-                        .testTag(COMPLETED_TASKS_TOGGLE)
-                ) {
-                    RowWithIcon(
-                        icon = {
-                            when {
-                                showCompleted -> Icon(LucideIcons.ChevronDown, null)
-                                else -> Icon(LucideIcons.ChevronRight, null)
-                            }
-                        }
-                    ) {
-                        Text(
-                            stringResource(Res.string.task_list_pane_completed_section_title_with_count, taskList.completedTasks.size),
-                            modifier = Modifier.testTag(COMPLETED_TASKS_TOGGLE_LABEL),
-                            style = MaterialTheme.typography.titleSmall
-                        )
-                    }
-                }
-            }
-        }
-
-        if (showCompleted) {
-            items(taskList.completedTasks, key = { it.id.value }) { task ->
-                CompletedTaskRow(
-                    task,
-                    onAction = { action ->
-                        when (action) {
-                            TaskAction.ToggleCompletion -> onToggleCompletionState(task)
-                            TaskAction.Edit -> onEditTask(task)
-                            TaskAction.UpdateDueDate -> onUpdateDueDate(task)
-                            TaskAction.Delete -> onDeleteTask(task)
-                            else -> Unit
-                        }
-                    },
-                )
-            }
-        }
-    }
-}
-
-private val DateRange.key: String
-    get() = when (this) {
-        is DateRange.Overdue -> "overdue${numberOfDays}"
-        is DateRange.Today -> "today"
-        is DateRange.Later -> "later${numberOfDays}"
-        DateRange.None -> "none"
-    }
-
